% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{activate_igraph}
\alias{activate_igraph}
\title{Use PhyloPics with igraph}
\usage{
activate_igraph()
}
\description{
This function adds a new vertex shape ("phylopic") to igraph that allows you
to use PhyloPics in your igraph plots. It sets up the necessary parameters
(see Details below) and registers the shape with igraph.
}
\details{
Additional details...
}
\examples{
\dontrun{
# load igraph and activate the connection with rphylopic
library(igraph)
activate_igraph()
# create a simple graph
g <- make_ring(10)
# plot the graph with phylopics
plot(g, vertex.shape = "phylopic", vertex.color = rainbow(vcount(g)),
     vertex.name = "Gorilla")
}
}
