\name{add_phylopic}
\alias{add_phylopic}
\title{Input an image and create a ggplot2 layer to add to an existing plot}
\usage{
  add_phylopic(img)
}
\arguments{
  \item{img}{A png object, e.g, from using
  \code{get_image}}
}
\description{
  Input an image and create a ggplot2 layer to add to an
  existing plot
}

