% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.r
\name{citationtonumber}
\alias{citationtonumber}
\title{Convert citation null to number 1}
\usage{
citationtonumber(x)
}
\arguments{
\item{x}{Input thing}
}
\description{
Convert citation null to number 1
}
\keyword{internal}

