% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylopic.R
\name{clear_phylopic_cache}
\alias{clear_phylopic_cache}
\title{Clear the rphylopic cache}
\usage{
clear_phylopic_cache()
}
\value{
No return value, called for side effects.
}
\description{
Clears both the HTTP response cache \code{\link[httpcache:clearCache]{httpcache::clearCache()}} and the parsed
image object cache used by \code{\link[=get_phylopic]{get_phylopic()}}.
}
