% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_phylo_names}
\alias{make_phylo_names}
\title{Make phylogeny with Phylopic images}
\usage{
make_phylo_names(query, size = size, phytheme = theme_phylo_blank2(),
  usenames = FALSE)
}
\arguments{
\item{input}{Object from get_image function}

\item{phytheme}{The ggplot2 style layout. You can write your own theme and pass
it in to the function. See example below.}
}
\description{
Make phylogeny with Phylopic images
}
\examples{
\dontrun{
# Attach taxon names to the figure
make_phylo_names(query="Homo sapiens", size="128", usenames=TRUE)
}
}

