% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_parallel.R
\name{message_parallel}
\alias{message_parallel}
\title{Send messages to console even from within parallel processes}
\usage{
message_parallel(...)
}
\value{
A message
}
\description{
Send messages to console even from within parallel processes
}
\keyword{internal}
