% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phylopic_count.r
\name{phylopic_count}
\alias{phylopic_count}
\title{Get the total number of PhyloPic images.}
\usage{
phylopic_count()
}
\description{
Get the total number of PhyloPic images.
}
\examples{
\dontrun{
phylopic_count()
}
}
\keyword{internal}

