% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{plot.phylopic}
\alias{plot.phylopic}
\title{Plot a raster representation of a PhyloPic silhouette}
\usage{
\method{plot}{phylopic}(img, ...)
}
\arguments{
\item{img}{A png array object, e.g., from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{...}{Other arguments passed on to \code{\link[grid:grid.raster]{grid::grid.raster()}}.}
}
\description{
Plot a raster representation of a PhyloPic silhouette
}
