% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{plot.Picture}
\alias{plot.Picture}
\title{Plot a vector representation of a PhyloPic silhouette}
\usage{
\method{plot}{Picture}(img, ...)
}
\arguments{
\item{img}{A \link[grImport2:Picture-class]{Picture} object, e.g., from using
\code{\link[=get_phylopic]{get_phylopic()}}.}

\item{...}{Other arguments passed on to \code{\link[grImport2:grid.picture]{grImport2::grid.picture()}}.}
}
\description{
Plot a vector representation of a PhyloPic silhouette
}
