% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{plot.Picture}
\alias{plot.Picture}
\alias{plot_phylopic}
\alias{plot.phylopic}
\title{Preview a PhyloPic silhouette}
\usage{
\method{plot}{Picture}(x, ...)

\method{plot}{phylopic}(x, ...)
}
\arguments{
\item{x}{A \link[grImport2:Picture-class]{Picture} or png array object, e.g.,
from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{...}{Other arguments passed on to \code{\link[grImport2:grid.picture]{grImport2::grid.picture()}} or
\code{\link[grid:grid.raster]{grid::grid.raster()}}.}
}
\description{
Preview a raster or vector representation of a PhyloPic silhouette. This will
plot the silhouette on a new page in your default plotting device.
}
