% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphylopic-package.R
\name{plot_phylopic_base}
\alias{plot_phylopic_base}
\title{This function is defunct.}
\usage{
plot_phylopic_base(...)
}
\description{
This function is defunct.
}
\note{
see \code{\link[=add_phylopic_base]{add_phylopic_base()}}
}
\keyword{internal}
