% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_phylopic.R
\name{resolve_phylopic}
\alias{resolve_phylopic}
\title{Resolve a name from another database to an image in the PhyloPic database}
\usage{
resolve_phylopic(
  name,
  api = "paleobiodb.org",
  hierarchy = FALSE,
  n = 1,
  filter = NULL
)
}
\arguments{
\item{name}{\code{character}. A full or partial taxonomic name to be matched
to the specified \code{api}.}

\item{api}{\code{character}. One of}

\item{hierarchy}{\code{logical}.}

\item{n}{\code{numeric}. How many uuids should be returned? Depending on
the requested \code{name}, multiple silhouettes might exist. If \code{n} exceeds
the number of available images, all available uuids will be returned.
This argument defaults to 1.}

\item{filter}{\code{character}. Filter uuid(s) by usage license. Use "by" to
limit results to image uuids which do not require attribution, "nc" for
image uuids which allow commercial usage, and "sa" for image uuids without
a ShareAlike clause. The user can also combine these filters as a vector.}
}
\value{
If \code{hierarchy} is \code{FALSE}, \code{n} valid PhyloPic image uuids for the
taxonomic node that most closely matches the supplied \code{name} as it relates
to the specified \code{api}. If \code{hierarchy} is \code{TRUE}, a list where the names
are the the taxonomic hierarchy for \code{name} as reported by the specified
\code{api} and the values are \code{n} resolved PhyloPic uuids for each taxonomic
entity (if they exist).
}
\description{
Resolve a name from another database to an image in the PhyloPic database
}
