% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{rotate_phylopic}
\alias{rotate_phylopic}
\title{Rotate a PhyloPic silhouette}
\usage{
rotate_phylopic(img, angle = 90)
}
\arguments{
\item{img}{A \link[grImport2:Picture-class]{Picture} or png array object, e.g.,
from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{angle}{\code{numeric}. The number of degrees to rotate the silhouette
clockwise.}
}
\value{
A \link[grImport2:Picture-class]{Picture} or png array object (matching
the type of \code{img})
}
\description{
The picture can be a \link[grImport2:Picture-class]{Picture} or png array object,
e.g., from using \code{\link[=get_phylopic]{get_phylopic()}}. Note that png array objects can only be
rotated by multiples of 90 degrees.
}
\seealso{
Other transformations:
\code{\link[=flip_phylopic]{flip_phylopic()}},
\code{\link[=recolor_phylopic]{recolor_phylopic()}}
}
\concept{transformations}
