% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphylopic-package.R
\docType{package}
\name{rphylopic}
\alias{rphylopic}
\alias{rphylopic-package}
\title{rphylopic}
\description{
Get Silhouettes of Organisms from Phylopic
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

David Miller \email{dave@ninepointeightone.net}
}
