% Generated by roxygen2 (4.0.1): do not edit by hand
\name{search_text}
\alias{search_text}
\title{Text search for uuids}
\usage{
search_text(text, options)
}
\arguments{
\item{text}{Search string, see examples}

\item{options}{See here for options}
}
\description{
These aren't necessarily ones with images though. See example
}
\examples{
\dontrun{
search_text(text = "Homo sapiens", options = "names")
}
}

