% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ubio.R
\name{ubio}
\alias{ubio}
\alias{ubio_get}
\title{Perform actions with uBio data.}
\usage{
ubio_get(namebankID, options = NULL, ...)
}
\arguments{
\item{namebankID}{The identifier for a name in uBio Namebank.}

\item{options}{(character) One or more of citationStart, html, namebankID, root, string,
type, uid, uri, and/or votes}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Perform actions with uBio data.
}
\details{
There is only one function for working with uBio data right now, \code{ubio_get()}

Options for the \code{options} parameter: Same as those for \code{name_*()} functions.
}
\examples{
\dontrun{
# Retrieves information on a set of taxonomic names.
ubio_get(109086)
ubio_get(109086, options=c('names','string'))
}
}

