% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_phylopic}
\alias{add_phylopic}
\title{Input an image and create a ggplot2 layer to add to an existing plot}
\usage{
add_phylopic(img)
}
\arguments{
\item{img}{A png object, e.g, from using \code{get_image}}
}
\description{
Input an image and create a ggplot2 layer to add to an existing plot
}
\examples{
\dontrun{
# Put a silhouette behind a plot
library(ggplot2)
img <- get_image("27356f15-3cf8-47e8-ab41-71c6260b2724", size = "512")[[1]]
qplot(x=Sepal.Length, y=Sepal.Width, data=iris, geom="point") + add_phylopic(img)
}
}

