% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_phylopic_tree.R
\name{add_phylopic_tree}
\alias{add_phylopic_tree}
\title{Add PhyloPics to a plotted phylogenetic tree}
\usage{
add_phylopic_tree(
  tree,
  tip,
  img = NULL,
  name = if (is.null(img) && is.null(uuid)) tip else NULL,
  uuid = NULL,
  width,
  relWidth = 0.06,
  padding = 1/200,
  hjust = 0,
  ...
)
}
\arguments{
\item{tree}{The phylogenetic tree object of class \code{phylo} on which to add
the silhouette.}

\item{tip}{The tip labels against which to add the silhouettes.}

\item{img}{A \link[grImport2:Picture-class]{Picture} or png array object, e.g.,
from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{name}{\code{character}. A taxonomic name to be passed to \code{\link[=get_uuid]{get_uuid()}}.}

\item{uuid}{\code{character}. A valid uuid for a PhyloPic silhouette (such as
that returned by \code{\link[=get_uuid]{get_uuid()}} or \code{\link[=pick_phylopic]{pick_phylopic()}}).}

\item{width}{\code{numeric}. Width of the silhouette in coordinate space. If
"NA", the default, and \code{height} is "NA", the silhouette will be as large as
fits in the plot area. If "NA", the default, and \code{height} is specified, the
width is determined by the aspect ratio of the original image. One or both
of \code{height} and \code{width} must be "NA".}

\item{relWidth}{The width of each silhouette relative to the plotting area.}

\item{padding}{Distance to inset each silhouette from the right edge of the
plotting area, relative to the size of the plotting area.
Negative values offset to the right.}

\item{hjust}{\code{numeric}. A numeric vector between 0 and 1 specifying
horizontal justification (left = 0, center = 0.5, right = 1). Note that,
due to the enforcement of the silhouette's aspect ratio, there may be
unexpected behavior due to interactions between the aspect ratio of the
plot and the aspect ratio of the silhouette.}

\item{\dots}{Further arguments to pass to \code{add_phylopic_base()}.}
}
\description{
Specify existing images, taxonomic names, or PhyloPic uuids to add PhyloPic
silhouettes alongside the associated leaves of a phylogenetic tree that has
been plotted in the active graphics device.
}
\examples{
\dontrun{
 # Load the ape library to work with phylogenetic trees
library("ape")

# Read a phylogenetic tree
tree <- ape::read.tree(text = "(cat, (dog, mouse));")

# Set a large right margin to accommodate the silhouettes
par(mar = c(1, 1, 1, 10))

# Plot the tree
plot(tree)

# Add a PhyloPic silhouette of a cat to the tree
add_phylopic_tree(
  tree, # Must be the tree that was plotted
  "cat", # Which leaf should the silhouette be plotted against?
  uuid = "23cd6aa4-9587-4a2e-8e26-de42885004c9", # Silhouette to plot
  relWidth = 0.2,
  fill = "brown"
)
}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
