% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{flip_phylopic}
\alias{flip_phylopic}
\title{Flip a PhyloPic silhouette along its horizontal and/or vertical axis}
\usage{
flip_phylopic(img, horizontal = TRUE, vertical = FALSE)
}
\arguments{
\item{img}{A \link[grImport2:Picture-class]{Picture} or png array object, e.g.,
from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{horizontal}{\code{logical}. Should the silhouette be flipped
horizontally?}

\item{vertical}{\code{logical}. Should the silhouette be flipped vertically?}
}
\value{
A \link[grImport2:Picture-class]{Picture} or png array object (matching
the type of \code{img})
}
\description{
The picture can be a \link[grImport2:Picture-class]{Picture} or png array object,
e.g., from using \code{\link[=get_phylopic]{get_phylopic()}}.
}
\seealso{
Other transformations:
\code{\link[=recolor_phylopic]{recolor_phylopic()}},
\code{\link[=rotate_phylopic]{rotate_phylopic()}}
}
\concept{transformations}
