% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_images.R
\name{gather_images}
\alias{gather_images}
\title{Gather 1 valid image per species}
\source{
\href{https://github.com/GuangchuangYu/ggimage/blob/master/R/geom_phylopic.R}{
Related function: \code{ggimage::geom_phylopic}}
}
\usage{
gather_images(
  species,
  options = c("namebankID", "names", "string"),
  include_image_data = FALSE,
  mc.cores = 1,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{species}{Species list.}

\item{options}{(character) One or more of citationStart, html, namebankID, root, string,
type, uid, uri, and/or votes}

\item{include_image_data}{Include the image data itself
(not just the image UID) in the results (default: \code{FALSE}).}

\item{mc.cores}{Accelerate multiple species queries by parallelising
across multiple cores (default: \code{1}).}

\item{verbose}{Print messages (default: \code{TRUE}).}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
data.frame with:
\itemize{
\item{species : }{Species name.}
\item{uid : }{Species UID.}
\item{namebankID : }{\href{http://ubio.org/}{uBio} species ID.}
\item{string : }{Standardised species name.}
\item{picid : }{Image UID.}
}
}
\description{
There are multiple images per species, and not all image UIDs are valid.
This function makes searching for a valid image for each species much easier.
Searches for the best name match per \code{species} and
iteratively tries image UIDs until a viable image is found.
}
\details{
Contributed by \href{https://github.com/bschilder}{Brian M. Schilder}.
}
\examples{
 
species <- c("Mus_musculus","Pan_troglodytes","Homo_sapiens")
res <- rphylopic::gather_images(species=species)
}
