% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_images.R
\name{gather_images}
\alias{gather_images}
\title{Gather 1 image per species}
\usage{
gather_images(
  species,
  options = c("namebankID", "names", "string"),
  include_image_data = FALSE,
  ...
)
}
\arguments{
\item{species}{Species list.}

\item{options}{(character) One or more of citationStart, html, namebankID, root, string,
type, uid, uri, and/or votes}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
There are multiple images per species, and not all pciture UIDs are valid.
This function makes searching for a valid image for each species much easier.
Searches for the best name match per \code{species} and
iteratively tries image UIDs until a viable PNG is found.
}
\examples{
 
species <- c("Mus_musculus","Pan_troglodytes","Homo_sapiens")
res <- gather_images(species=species)
}
