% Generated by roxygen2 (4.0.2): do not edit by hand
\name{image}
\alias{image}
\alias{image_count}
\alias{image_get}
\alias{image_list}
\alias{image_timerange}
\title{Perform actions with images.}
\usage{
image_get(uuid, options = NULL, ...)

image_list(start = 1, length = 10, options = NULL, ...)

image_timerange(timestamp = "modified", from = NULL, to = NULL,
  options = NULL, ...)

image_count(...)
}
\description{
Perform actions with images.
}
\details{
I'm not adding methods for modifying images, including add, edit, updated, delete, and
transfer, because I can't imagine doing those things from R. Am I wrong?

Note: uid is always returned
}
\examples{
\dontrun{
# Get info on an image
id <- "9fae30cd-fb59-4a81-a39c-e1826a35f612"
image_get(uuid = id)
image_get(uuid = id, options=c('credit','pngFiles','taxa','canonicalName','string','uri','type'))
image_get(uuid = id, options=c('credit','licenseURL','pngFiles','submitted','submitter',
   'svgFile','taxa','canonicalName','string','uri','type','citationStart'))

# Count images in Phylopic database
image_count()
library('httr')
image_count(config = verbose())

# Lists images in chronological order, from most to least recently modified
image_list(start=1, length=10)
image_list(start=1, length=10, options="taxa")

# Lists images within a given time range, from most to least recent
image_timerange(from="2013-05-11")
image_timerange(from="2013-05-11", to="2013-05-12")
image_timerange(from="2013-05-11", to="2013-05-12", options='credit')
}
}

