% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image}
\alias{image}
\alias{image_get}
\alias{image_list}
\alias{image_timerange}
\alias{image_count}
\alias{image_data}
\title{Perform actions with images.}
\usage{
image_get(uuid, options = NULL, ...)

image_list(page = 0, options = NULL, ...)

image_timerange(options = NULL, ...)

image_count(options = NULL, ...)

image_data(uuid, size = "vector", ...)
}
\arguments{
\item{uuid}{One or more name UUIDs.}

\item{options}{(character) One or more of citationStart, html, namebankID,
root, string, type, uid, uri, and/or votes.}

\item{...}{curl options passed on to \link[httr:GET]{httr::GET}}

\item{page}{The page of the results to return.}

\item{size}{Height of the image, one of 64, 128, 256, 512, 1024, "thumb", or
"icon"}
}
\description{
Perform actions with images.
}
\details{
I'm not adding methods for modifying images, including add, edit,
updated, delete, and transfer, because I can't imagine doing those things
from R. Am I wrong?

Note: uid is always returned
}
\examples{
\dontrun{
# Get info on an image
uuid <- "9fae30cd-fb59-4a81-a39c-e1826a35f612"
image_get(uuid = uuid)

# Count images in Phylopic database
image_count()

# Lists images in chronological order, from most to least recently modified
image_list()
image_list(page = 10)

# Return the timerange of images in phylopic
image_timerange()

# Get data for an image
image_data(uuid, size = "vector") # vector format
image_data(uuid, size = "512") # raster format
}
}
