% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{image}
\alias{image}
\alias{get_phylopic}
\title{Perform actions with images.}
\usage{
get_phylopic(uuid, size = "vector")
}
\arguments{
\item{uuid}{One or more name UUIDs.}

\item{size}{Height of the image, one of 64, 128, 256, 512, 1024, "thumb", or
"icon"}
}
\description{
Perform actions with images.
}
\details{
I'm not adding methods for modifying images, including add, edit,
updated, delete, and transfer, because I can't imagine doing those things
from R. Am I wrong?

Note: uid is always returned
}
\examples{
\dontrun{
# Get info on an image
uuid <- "9fae30cd-fb59-4a81-a39c-e1826a35f612"
image_get(uuid = uuid)

# Count images in Phylopic database
image_count()

# Lists images in chronological order, from most to least recently modified
image_list()
image_list(page = 10)

# Return the timerange of images in phylopic
image_timerange()

# Get data for an image
get_phylopic(uuid, size = "vector") # vector format
get_phylopic(uuid, size = "512") # raster format
}
}
