% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_phylo}
\alias{make_phylo}
\title{Make phylogeny with Phylopic images}
\usage{
make_phylo(pngobj, phytheme = theme_phylo_blank2())
}
\arguments{
\item{pngobj}{Object from get_image function}

\item{phytheme}{The ggplot2 style layout. You can write your own theme and pass
it in to the function. See example below.}
}
\description{
Make phylogeny with Phylopic images
}
\examples{
\dontrun{
toget <- c("27356f15-3cf8-47e8-ab41-71c6260b2724", "bd88f674-6976-4cb2-a46e-e6a12a8ba463", "e547cd01-7dd1-495b-8239-52cf9971a609", "9c6af553-390c-4bdd-baeb-6992cbc540b1", "e547cd01-7dd1-495b-8239-52cf9971a609", "bd88f674-6976-4cb2-a46e-e6a12a8ba463")
myobjs <- get_image(uuids = toget, size = "thumb")
make_phylo(pngobj=myobjs)

# Settingn phytheme to NULL uses ggphylo's default theme
make_phylo(pngobj=myobjs, phytheme=NULL)

# Make your own theme
mytheme <- theme_phylo_blank2() + ggplot2::theme(panel.background=element_rect(fill="lightblue"))
make_phylo(pngobj=myobjs, phytheme=mytheme)

# The use case starting from a text search
searchres <- search_text(text = "Homo sapiens", options = "names")
output <- search_images(uuid=searchres, options=c("pngFiles", "credit", "canonicalName"))
myobjs <- get_image(uuids = output, size = "128")
make_phylo(pngobj=myobjs)
}
}

