% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylopic_utils.R
\name{recolor_phylopic}
\alias{recolor_phylopic}
\title{Recolor a phylopic image}
\usage{
recolor_phylopic(img, alpha = 1, color = NULL)
}
\arguments{
\item{img}{A \link[grImport2:Picture-class]{Picture} or png array object, e.g.,
from using \code{\link[=get_phylopic]{get_phylopic()}}.}

\item{alpha}{A value between 0 and 1, specifying the opacity of the
silhouette.}

\item{color}{Color to plot the silhouette in. If NULL, the color is not
changed.}
}
\value{
A \link[grImport2:Picture-class]{Picture} or png array object (matching
the type of \code{img})
}
\description{
Function to recolour and change alpha levels of a phylopic image.
}
