% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rphylopic-package.R
\docType{package}
\name{rphylopic-package}
\alias{rphylopic}
\alias{rphylopic-package}
\title{rphylopic: Get Silhouettes of Organisms from PhyloPic}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Work with the PhyloPic Web Service (\url{http://api-docs.phylopic.org/v2/}) to fetch silhouette images of organisms. Includes functions for adding silhouettes to both base R plots and ggplot2 plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rphylopic.palaeoverse.org}
  \item \url{https://github.com/palaeoverse/rphylopic}
  \item \url{https://palaeoverse.org}
  \item Report bugs at \url{https://github.com/palaeoverse/rphylopic/issues}
}

}
\author{
\strong{Maintainer}: William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})

Authors:
\itemize{
  \item William Gearty \email{willgearty@gmail.com} (\href{https://orcid.org/0000-0003-0076-3262}{ORCID})
  \item Lewis A. Jones \email{lewis.jones@ucl.ac.uk} (\href{https://orcid.org/0000-0003-3902-8986}{ORCID})
}

Other contributors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [contributor]
  \item Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [contributor]
  \item David L. Miller \email{dave@ninepointeightone.net} (\href{https://orcid.org/0000-0002-9640-6755}{ORCID}) [contributor]
}

}
\keyword{internal}
