% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rphylopic-package.R
\docType{package}
\name{rphylopic}
\alias{rphylopic}
\alias{rphylopic-package}
\title{rphylopic}
\description{
rphylopic
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

David Miller \email{dave@ninepointeightone.net}
}

