% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_phylopic.R
\name{save_phylopic}
\alias{save_phylopic}
\title{Save a PhyloPic image}
\usage{
save_phylopic(img = NULL, path = NULL, width = NULL, height = NULL)
}
\arguments{
\item{img}{\code{Picture?}.}

\item{path}{\code{character}.}

\item{width}{\code{numeric}.}

\item{height}{\code{numeric}.}
}
\description{
\code{save_phylopic} is a convenient function for saving a PhyloPic image.
Available formats for saving include svg, pdf, png, ps, and eps.
}
