% Generated by roxygen2 (4.0.1): do not edit by hand
\name{search_images}
\alias{search_images}
\title{Search for images for the taxa}
\usage{
search_images(uuid, subtaxa = NULL, supertaxa = NULL, options = NULL,
  cleanoutput = TRUE, ...)
}
\arguments{
\item{uuid}{The UUID of the taxonomic name.}

\item{subtaxa}{If set to "true", includes subtaxa in the search.}

\item{supertaxa}{If not set to "false", includes supertaxa in the search.}

\item{options}{Space-separated list of options for the result value.}

\item{cleanoutput}{If TRUE, remove elements with no data.}

\item{...}{Further args passed on to GET. See examples.}
}
\description{
Search for images for the taxa
}
\examples{
\dontrun{
someuuids <- search_text(text = "Homo sapiens", options = "names")
search_images(uuid=someuuids[[4]], options=c("pngFiles", "credit", "canonicalName"))

# all of them
search_images(uuid=someuuids, options=c("pngFiles", "credit", "canonicalName"))
}
}

