\name{search_text}
\alias{search_text}
\title{Text search for uuids}
\usage{
  search_text(text, options)
}
\arguments{
  \item{text}{}

  \item{options}{}
}
\description{
  These aren't necessarily ones with images though. See
  example
}
\examples{
search_text(text = "Homo sapiens", options = "names")
}

